compilation warnings

+++ storage/maria/trnman.c
@@ -390,7 +390,7 @@ my_bool trnman_end_trn(TRN *trn, my_bool commit)
   TRN *free_me= 0;
   LF_PINS *pins= trn->pins;
   DBUG_ENTER("trnman_end_trn");
-  DBUG_PRINT("enter", ("trn=0x%lx commit=%d", trn, commit));
+  DBUG_PRINT("enter", ("trn=0x%lx commit=%d", (ulong) trn, commit));
 
   DBUG_ASSERT(trn->rec_lsn == 0);
   /* if a rollback, all UNDO records should have been executed */
